Addressed race condition with a vitest spy by moving function to an imported file for testing
